requirements.txt
pandas~=1.2.4
termcolor==1.1.0
scikit-learn~=0.24.2
numpy=1.21.2
#python-louvain==0.11
networkx==2.5.1
scipy~=1.6.3
opencv-python~=4.5.2.52
matplotlib~=3.4.2
lap~=0.4.0
scikit-image~=0.18.1
filterpy~=1.4.5
python-dateutil~=2.8.1
torch~=1.8.1
Pillow~=8.2.0
six~=1.16.0
torchvision~=0.9.1
tqdm~=4.61.0
numpy~=1.20.3
colorama~=0.4.4
joblib~=1.0.1
PyYAML~=5.4.1
wandb~=0.11.2
seaborn~=0.11.2
setuptools~=57.0.0
azure~=4.0.0
msrest~=0.6.21
StringIO
